+++ src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/utils/Annotations.kt
@@ -1,0 +1,6 @@
+package com.github.insanusmokrassar.TelegramBotAPI.utils
+
+import kotlin.Experimental.*
+
+@Experimental(Level.WARNING)
+annotation class PreviewFeature