+++ CHANGELOG.md
@@ -68,6 +68,15 @@ mistake - don't hesitate to say this.**
 * Version updates:
     * Ktor `1.2.6` -> `1.3.0`
 
+### 0.22.1 MediaContent#asInputMedia
+
+* All `MediaContent` instances now can create their `InputMedia` analog
+* New annotation `PreviewFeature` was added to mark new thing as preview for the time
+while they can work incorrectly
+* Added links utils:
+    * `makeLinkToMessage` have two signatures - for direct creating using username and for abstract creating using
+    chat id
+
 ## 0.21.0 TelegramBotAPI 4.5
 
 * _**All `MessageEntity`'es now are replaced with `TextPart`**_